Conflicts:
	module/order/app/models/order/OrderShippingInfo.java
	module/order/app/models/order/OrderStatus.java
	operate/business/app/views/DownloadOrderShippingInfos/index.html

+++ resale/home/app/controllers/BatchExportCoupons.java
@@ -116,7 +116,7 @@ public class BatchExportCoupons extends Controller {
         batchCoupons.save();
         for (int i = 0; i < count; i++) {
             Order order = Order.createConsumeOrder(resaler.getId(), AccountType.RESALER);
-            int number = 1;
+            Long number = 1l;
             String phone = null;
             BigDecimal resalerPrice = goods.getResalePrice();
             order.addOrderItem(goods, number, phone,
